Check and log all lagging consumers.
The call to gtg can return immediately when a single lagging consumer is
found, so checking the remaining consumers and logging is passed off to
a goroutine.

This doesn't change the functional response of the lagcheck itself, but
it does record for posterity the lagging consumers in the logs, so that
we can determine if in any given event there is a single slow consumer,
or a system-wide problem.